Correct label shape (#4)
* Update to accept appropriate label shape

In Keras, labels have an extra dimension of size 1 when using sparse
softmax instead of one-hot vectors. To keep consistent, the
AdaptiveSoftmaxProduceLogits now expects labels to be provided with
the same dimensions. This moves the layer closer to being a drop in
replacement for a Dense layer using sparse_categorical_crossentropy.

* Add docker support for development

This includes a few helper scripts for building the docker
container, starting a bash shell, etc. which are in the scripts
subdirectory.

* Fix output shape for DifferentiatedSoftmaxProduceLogits

* Update text8 benchmark

The benchmark now handles the correct label shapes. Also some
additional changes were made to bring it closer in line with
facebook's benchmark written in torch:

    https://github.com/facebookresearch/adaptive-softmax

There are still some variations. The adagrad implementation is
different with respect to how the provided epsilon value is used,
and their norm clipping does not appear to include the entire
gradient vector.

* Update performance comparison

* Add tests script to run tests using docker